ci(l2): always run TDX lints (#2912)
**Motivation**

In #2867 a change broke TDX, but this wasn't caught by the CI because
the TDX workflow isn't executed on PRs that do not change TDX-related
files.

**Description**

This moves the lint task with the other prover lints, so that it runs on
every PR.

The TDX test is still only executed selectively.