ci(l2): add github workflow for releases (#2853)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We need an automated way to make releases for the L2.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Added a new Github workflow that cross-compiles the L2 components and
makes a new release with the binaries.
To trigger the workflow, a tag with name `l2_vX.Y.Z` has to be created.
A changelog will be auto-generated based on the last tag before this
one. Binaries are compiled for linux (x86-64 & arm64) and macos (arm64)

<!-- Link to issues: Resolves #111, Resolves #222 -->