refactor(levm): simplified shift functions to improve readability and performance (#2933)
**Motivation**

Make the VM code that handles bitwise shifts both easier to read and
better performing.

**Description**

This PR removes the checked shift right and shift left functions. The
EVM spec for the SHL, SHR and SAR opcodes doesn't require checking for
overflow or underflow; and doing the shifts by multiplying or dividing
significantly degraded performance. This PR changes those for the bit
shift operator.

Closes #2930