feat(l1): compare levm logs against revm's (#2932)
**Motivation**

When running EFTests we check that the post state root is the
appropriate but we weren't checking that the logs were correct.

**Description**

This pr adds code for comparing levm and revm logs. First the hashes of
a rlp encoding of both are compared and in case of a mismatch a more
detailed comparison is made.

Closes #2594

---------

Co-authored-by: SDartayet <sofiadartayet@gmail.com>