feat(l1): improve ancestor retrieval for canonical blocks (#3005)
**Motivation**
During Evm execution, we might need to look for a block hash in the
state. Previously we encountered bugs as we were looking up the
canonical block hash for the given block number when we should be
searching for the block hash in the current block's (the one the
EvmState was built on) ancestors. While this was already fixed by #2965,
it made block hash lookup much slower.
In our most common use cases (fork choice updates, new payloads, full
sync). The current block will be a canonical one, so we know all of its
ancestors will also be canonical, and there is no reason to go through
all of its ancestors. So this PR now checks if the current block is
canonical, if it is, it searches for the block hash in the canonical
block hashes, and if it is not, it searches through its ancestors
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* `StoreVmDatabase::get_block_hash` now searches for the target hash in
canonical hashes if its current block (the one set at creation) is
canonical
* Add sync versions of `StoreEngine` methods `get_block_number` &
`get_canonical_block_hash`
* Add `Store` method `is_canonical_sync` to check weather a block hash
is canonical synchronously
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2990