fix(l1): cache block hashes when adding blocks in batch (#3034)
**Motivation**
When adding blocks in batches, transactions in blocks may need to read
the hash of a previous block in the batch. This poses an issue, as we
add all blocks after the full batch has been executed ad we don't have
those hashes stored.
Previously, this issue was tackled by storing the block headers before
executing and storing the full block batch. This also required us to
make the fixes on #2831 (as we might have the headers of blocks we have
not yet executed during full sync).
This stopped being a suitable solution once #3022 was merged, as block
hashes are now searched using the block the state is based on's
ancestors. As the state is based on the first block in the batch's
parents, later blocks are not able to access the hashes of previous
blocks in the batch.
This PR aims to fix both issues by adding a `block_hash_cache` to the
`StoreVmDatabase` were we would add the hashes of all blocks in the
batch, and use it when fetching block hashes before looking at the
Store. This way we can access block hashes for previously executed
blocks in the batch without prematurely adding data from blocks yet to
be executed to the state. Allowing us to rollback fixes from #2831
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add `block_hash_cache` to `VmStoreDatabase`
* Cache block hashes of the full batch before executing in
`Blockchain::add_block_batch`
* No longer store block headers before execution during full sync
* Rollback fixes from #2831 (no longer needed due to the above change)
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number