refactor(core): make TrieDB::put a provided method (#3038)
**Motivation**

Remove redundant code.

**Description**

Simplify the TraitDB implementations by treating `put` as a special
case of `put_batch`, implementing it as a provided method in the
trait itself.

**Discussion**

The method itself is only used in tests, so it might be better to
change the tests to use the actual `put_batch` and just delete the
`put` method.