fix(l1): `are_block_headers_chained` (GetBlockHeaders validation) (#3049)
**Motivation**
The current `are_block_headers_chained` only works when the ordering is
`OldToNew` but fails if the ordering is `NewToOld`. As both orderings
are accepted by the request method both should be appropriately
validated.
This PR fixes this problem and also refactors the method to iterate
block headers in windows of 2 instead of relying on zipping iterators
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Use `BlockRequestOrdering` in `are_block_headers_chained`
* Simplify code of `are_block_headers_chained`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number