feat(l2): prove state diffs (#2819)
**Motivation**

The goal is that the prover includes the previously committed blob
versioned hash as a public input, and checks that it's correct. This is
the last big feature missing from the prover. After merging this PR, all
fields in the L1 block commitment will be validated.

- The blob (encoded state diffs), its commitment and proof are created
by the L1 Committer.
- The L1 Committer will send the `commitBatch` transaction (EIP4844),
embedding the blob data. Then the contract will recover the blob
versioned hash with the `BLOBHASH` opcode and store it in the block
commitment.
- The prover will take theblob commitment and blob proof as private
inputs. The prover will compute the blob versioned hash, set it as
public input and do all necessary checks:
- First it will compute the state diffs from the account updates
resulted from the batch execution
- Then it will compute the valid blob by encoding the state diffs, and
verify the proof (proof that the commitment binds to the valid blob)
- After that, it computes the blob versioned hash from the blob
commitment and sets it as a public input, so now the L1 can verify that
the previously commited blob versioned hash is valid.

Verifying the blob proof in the zkvm guest and later verifying the zkvm
proof in the L1 is equivalent to calling the point evaluation precompile
in the L1, following the protocol as defined in `c-kzg` (the blob proof
is an opening over a challenge, which is defined using the blob
commitment and blob data).

We leverage `kzg-rs` to verify the blob proof. `kzg-rs` is a
zkvm-friendly replacement for `c-kzg`, without proving capabilities
(because a zkvm guest is only interested in verification, the proof can
be calculated in the host).

**Description**

- created `ethrex-l2-common` crate and refactored state diffs into there
so that both `ethrex-l2` and `zkvm-interface` can use them
- also refactored into `ethrex-l2-common` duplicated functions related
to withdrawals and deposits.
- L1 committer pushes blob bundles (blobs, its commitment and proof)
into a cache
- Proof coordinator retrieves a bundle, extracts the blob for the batch
to prove and sends the blob commitment and blob proof to the prover as a
private input.
- prover computes (as part of its program) the state diffs
- prover verifies the KZG blob proof (checks that the blob commitment
binds to the valid state diff) and computes the blob versioned hash from
the blob commitment, which then commits as a public input.
- removed perf_zkvm tests for convenience (had to do extra work to
create blobs for the test data, but we were planning removing these
tests anyway).
- changed `StateDiff` to use `BTreeMap` for data ordering and
deterministc encoding (necessary for deterministic commitments)
- created `L2Options` and moved the `validium` opt from the commiter to
there, because now the proof coordinator is also interested in that
option.
- now the blob versioned hash isn't passed by calldata during the
`commitBatch` transaction but it's recovered using the `BLOBLHASH`
opcode
- removed the `VALIDIUM` variable in the OnChainProposer contract

 
Closes #1100