feat(levm): add backup hook (#3069)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Add a new Backup Hook that is specifically for restoring state. It
allows us to do that cleanly in various parts of the code.

**Description**

- The goal is to make code cleaner and to utilize the backup hook in
`stateless_execute`. Before we were cloning the whole cache and it was
potentially expensive. With this solution we can seamlessly restore the
state 😄
- Improve integration with L2 so that changes made by a transaction that
doesn't fit the blob are smoothly reversed from the cache. We don't need
`execute_tx_l2` outside of the VM and our `execute` function in LEVM is
now cleaner.

Closes #3068